Allow for boolean OR attributes in model
Currently, there is a small deviation in this gem from the intended
behavior of GTFS. Only one of the "short_name" and "long_name"
attributes are required, not both. This adds another class variable to
capture this logic, by allowing you to specify groups of attributes,
one of which is required.

It's probably possible to jam this into the required_attrs class
variable, but I thought it was probably nicer to simply add another
category than worry about arrays vs. symbols in that variable.

This then also corrects the aforementioned behavior so that the Routes
model requires either the short_name or long_name, but not both (you
can still have both, however).